client: make max_std(err/out)_file_size double
File sizes should always be double, not int

+++ lib/cc_config.h
@@ -172,8 +172,8 @@ struct CC_CONFIG {
     int max_event_log_lines;
     int max_file_xfers;
     int max_file_xfers_per_project;
-    int max_stderr_file_size;
-    int max_stdout_file_size;
+    double max_stderr_file_size;
+    double max_stdout_file_size;
     int max_tasks_reported;
     int ncpus;
     bool no_alt_platform;
